requirements.txt
africastalking==1.2.4
aniso8601==9.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contextlib2==21.6.0
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.3
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
schema==0.7.5
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.10.0
gunicorn
pandas
python-dateutil
marshmallow
python-dateutil
openpyxl
xlsxwriter==1.4.5
boto3==1.17.100
faker>=8.0.0,<9.0.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
